chore: snack -> customview 변경
