+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/presentation/customview/CallSnackBar.kt
@@ -1,4 +1,4 @@
-package com.droidblossom.archive.presentation.snack
+package com.droidblossom.archive.presentation.customview
 
 data class CallSnackBar(
     // 타입, 아이디 등등 추가
