+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/presentation/ui/home/HomeFragment.kt
@@ -3,10 +3,8 @@ package com.droidblossom.archive.presentation.ui.home
 import android.graphics.Point
 import android.location.Location
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContentProviderCompat.requireContext
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.toPointF
 import androidx.fragment.app.viewModels
@@ -17,8 +15,8 @@ import com.droidblossom.archive.R
 import com.droidblossom.archive.databinding.FragmentHomeBinding
 import com.droidblossom.archive.domain.model.common.CapsuleMarker
 import com.droidblossom.archive.presentation.base.BaseFragment
-import com.droidblossom.archive.presentation.snack.HomeSnackBarBig
-import com.droidblossom.archive.presentation.snack.HomeSnackBarSmall
+import com.droidblossom.archive.presentation.customview.HomeSnackBarBig
+import com.droidblossom.archive.presentation.customview.HomeSnackBarSmall
 import com.droidblossom.archive.presentation.ui.home.createcapsule.CreateCapsuleActivity
 import com.droidblossom.archive.presentation.ui.home.dialog.CapsulePreviewDialogFragment
 import com.droidblossom.archive.util.CapsuleTypeUtils
@@ -35,11 +33,7 @@ import com.naver.maps.map.overlay.Overlay
 import com.naver.maps.map.overlay.OverlayImage
 import com.naver.maps.map.util.FusedLocationSource
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import kotlin.random.Random
 
 @AndroidEntryPoint
 class HomeFragment : BaseFragment<HomeViewModelImpl, FragmentHomeBinding>(R.layout.fragment_home),
