+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/util/MyFirebaseMessagingService.kt
@@ -12,7 +12,6 @@ import android.os.Build
 import android.util.Log
 import androidx.core.app.NotificationCompat
 import com.droidblossom.archive.R
-import com.droidblossom.archive.presentation.snack.CallSnackBar
 import com.droidblossom.archive.presentation.ui.MainActivity
 import com.google.firebase.messaging.FirebaseMessaging
 import com.google.firebase.messaging.FirebaseMessagingService
